requirements.txt
asgiref==3.5.2
dj-rest-auth==2.2.5
Django==3.0.8
django-allauth==0.51.0
django-environ==0.4.5
django-filter==21.1
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.0.4
PyJWT==1.7.1
mysqlclient==2.1.1
PyMySQL==1.0.2
pytz==2020.1
sqlparse==0.3.1